refactor trading momentum related components
- rename ProfitIndicator -> UpDownIndicator and improve API
- move TopTradesTable out of `components` since it has `frontend` dependencies
- co-locate momentum-related components in `lib/momentum`
- remove unused TopMomentumTile component
