+++ src/lib/components/index.ts
@@ -25,7 +25,6 @@ export { default as MenuItem } from './MenuItem.svelte';
 export { default as NavPanel } from './NavPanel.svelte';
 export { default as PageHeader } from './PageHeader.svelte';
 export { default as PageHeading } from './PageHeading.svelte';
-export { default as ProfitIndicator } from './ProfitIndicator.svelte';
 export { default as Section } from './Section.svelte';
 export { default as SmartContractWidget } from './SmartContractWidget.svelte';
 export { default as SocialIcon } from './SocialIcon.svelte';
@@ -34,5 +33,5 @@ export { default as SummaryBox } from './SummaryBox.svelte';
 export { default as Tabs } from './Tabs.svelte';
 export { default as TextInput } from './TextInput.svelte';
 export { default as Timestamp } from './Timestamp.svelte';
-export { default as TopTradesTable } from './TopTradesTable.svelte';
 export { default as TradingDataInfoRow } from './TradingDataInfoRow.svelte';
+export { default as UpDownIndicator } from './UpDownIndicator.svelte';
