+++ src/routes/+page.svelte
@@ -4,7 +4,8 @@
 <script lang="ts">
 	import type { PageData } from './$types';
 	import HomeHeroBanner from './home/HeroBanner.svelte';
-	import { BlogRoll, Button, Section, SummaryBox, TopTradesTable } from '$lib/components';
+	import { BlogRoll, Button, Section, SummaryBox } from '$lib/components';
+	import TopTradesTable from '$lib/momentum/TopTradesTable.svelte';
 	import { toggleSubscribeDialog } from '$lib/newsletter/controller';
 	import { sitelinksSearchBox } from '$lib/helpers/googleMeta';
